requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langgraph>=0.2.0
dashscope>=1.20.0
faiss-cpu>=1.8.0
python-dotenv>=1.0.0
rich>=13.0.0
